@pcrespov review: move to schemas
