+++ services/director-v2/tests/integration/test_computation_api.py
@@ -21,8 +21,8 @@ from pydantic.types import PositiveInt
 from simcore_postgres_database.models.comp_tasks import comp_tasks
 from simcore_postgres_database.models.projects import ProjectType, projects
 from simcore_postgres_database.models.users import UserRole, UserStatus, users
-from simcore_service_director_v2.models.domains.comp_tasks import ComputationTaskOut
 from simcore_service_director_v2.models.domains.projects import ProjectAtDB
+from simcore_service_director_v2.models.schemas.comp_tasks import ComputationTaskOut
 from sqlalchemy import literal_column
 from starlette import status
 from starlette.testclient import TestClient
